refactor(booking): extract getLocationLabel & test
